--- src/qibo/models/dbi/double_bracket_oracles.py
@@ -1,370 +1,0 @@
-from copy import deepcopy
-from enum import Enum, auto
-from functools import partial
-
-import hyperopt
-import numpy as np
-
-from qibo.config import raise_error
-from qibo.hamiltonians import Hamiltonian
-
-
-
-class DoubleBracketDiagonalAssociationType(Enum):
-    """Define the evolution generator of a variant of the double-bracket iterations."""
-    
-    dephasing = auto()
-    """Use dephasing for a canonical bracket."""
-    
-    prescribed = auto()
-    """Use some input diagonal matrix for each step: general diagonalization DBI"""
-
-    fixed = auto()
-    """Use same input diagonal matrix in each step: BHMM DBI""" 
-
-    optimization = auto()
-    """Perform optimization to find best diagonal operator""" 
-
-
-class EvolutionOracleType(Enum):  
-    text_strings = auto()
-    """If you only want to get a sequence of names of the oracle"""
-
-    numerical = auto()
-    """If you will work with exp(is_k J_k) as a numerical matrix"""
-
-    hamiltonian_simulation = auto()
-    """If you will use SymbolicHamiltonian"""
-
-
-class EvolutionOracle:
-    def __init__( h_generator: AbstractHamiltonian = None, 
-            name = None,
-            mode_evolution_oracle: EvolutionOracleType = EvolutionOracleType.text_strings ):
-       if mode_evolution_oracle is EvolutionOracleType.hamiltonian_simulation and if type(h_generator) is not SymbolicHamiltonian:
-            raise_error(TypeError, "If the evolution oracle mode will be to make Trotter-Suzuki decompositions then you must use the SymbolicHamiltonian generator")
-
-       self.h = h_generator
-       self.mode_evolution_oracle = mode_evolution_oracle
-       self.name = name
-
-    def __call__(self, t_duration: double = None):
-        """ Returns either the name or the circuit """
-        if t_duration is None:
-            return self.name
-        else:
-            return self.circuit( t_duration = t_duration )
-        
-    def circuit(self, t_duration: double = None):
-
-        if self.mode_evolution_oracle is EvolutionOracleType.text_strings:
-            return self.name + str(t_duration)
-        elif self.mode_evolution_oracle is EvolutionOracleType.numerical:
-            raise_error(NotImplementedError, "The following should add a big single gate and return a circuit")
-            return qibo.Circuit.add_custom_gate(self.h.exp(t_duration))
-        elif self.mode_evolution_oracle is EvolutionOracleType.hamiltonian_simulation:
-            return self.h.circuit(t_duration)
-        else:
-            raise_error(ValueError,
-                    f"You are using an EvolutionOracle type which is not yet supported.")
-    @property
-    def name(self):
-        return self.name
-
-#class DiagonalAssociation(EvolutionOracle):
-#
-#    def __init__( h_generator: AbstractHamiltonian = None, 
-#            name = None,
-#            mode_diagonal_association: DoubleBracketDiagonalAssociationType = DoubleBracketDiagonalAssociationType.dephasing,
-#            mode_evolution_oracle: EvolutionOracleType = EvolutionOracleType.text_strings
-#    ):        
-#        self.name = name
-#        self.mode_diagonal_association = mode_diagonal_association  
-#        self.mode_evolution_oracle = mode_evolution_oracle
-#    @property
-#    def name(self):
-#        return self.name
-#
-class DiagonalAssociationDephasing(EvolutionOracle):
-
-        def __init__(
-        self,
-        mode_evolution_oracle: EvolutionOracleType = EvolutionOracleType.text_strings
-    ):
-            super().__init__(mode_diagonal_association = DoubleBracketDiagonalAssociationType.dephasing, mode_evolution_oracle = mode_evolution_oracle)
-
-        def __call__(self, J_input: EvolutionHamiltonian, k_step_number: list = None, t_duration = None ):
-        if mode_evolution_oracle is EvolutionOracleType.text_strings:
-            if t_duration is None:
-                #iterate over all Z ops
-                return '\Delta(' + J_input.name + ')'
-            else:
-                return 'exp( i'+ str(t_duration) +'\Delta(' + J_input.name + ')'
-        elif mode_evolution_oracle is EvolutionOracleType.numerical:
-            if t_duration is None:
-                return J_input.h.diag()
-            else:
-                return J_input.diag().exp(t_duration) 
-        if mode_evolution_oracle is EvolutionOracleType.hamiltonian_simulation:
-            if t_duration is None:
-                #iterate over all Z ops
-                return sum Z @ J_input @ Z
-            else:
-                return sum Z @ J_input.circuit(t_duration) @Z
-
-class DiagonalAssociationFromList(EvolutionOracle):
-
-    def __init__(
-            self,
-            d_k_list: list = None,
-            mode_evolution_oracle: EvolutionOracleType = EvolutionOracleType.text_strings
-        ):
-        super().__init__(mode_diagonal_association = DoubleBracketDiagonalAssociationType.prescribed, mode_evolution_oracle = mode_evolution_oracle)
-        self.d_k_list = d_k_list
-
-    def __call__(self, k_step_number: list = None, t_duration = None ):
-        
-        if mode_evolution_oracle is EvolutionOracleType.text_strings:
-            if t_duration is None:
-                return 'D_' + str(k_step_number)
-            else:
-                return 'exp( i'+ str(t_duration) +'D_k'
-        elif mode_evolution_oracle is EvolutionOracleType.numerical:
-            if t_duration is None:
-                return d_k_list[k_step_number]
-            else:
-                return  d_k_list[k_step_number].exp(t_duration)
-        if mode_evolution_oracle is EvolutionOracleType.hamiltonian_simulation:
-            if t_duration is None:
-                raise_error(ValueError, f"In the hamiltonian_simulation mode you need to work with evolution operators so please specify a time.")
-            else:
-                return d_k_list[k_step_number.circuit(t_duration)
-
-class DiagonalAssociationFromOptimization(EvolutionOracle):
-
-    def __init__(
-    self,
-    loss_function: None,
-    mode: DoubleBracketDiagonalAssociationType = DoubleBracketDiagonalAssociationType.canonical,
-    mode_evolution_oracle: EvolutionOracleType = EvolutionOracleType.text_strings
-):
-        self.loss = loss_function
-
-    def __call__(self, h: AbstractHamiltonian, k_step_number: list = None, t_duration = None ):
-        if self.mode_evolution_oracle is EvolutionOracleType.text_strings:
-            if t_duration is None:
-                return 'Optimize $\mu$ D_' + str(k_step_number)
-            else:
-                return 'Optimize  $\mu$ exp( i'+ str(t_duration) +'D_k'
-        elif self.mode_evolution_oracle is EvolutionOracleType.numerical:
-            if t_duration is None:
-                raise_error(TypeError, "Not implemented")
-                return 0 
-            else:
-                raise_error(TypeError, "Not implemented")
-                return 0
-        if self.mode_evolution_oracle is EvolutionOracleType.hamiltonian_simulation:
-            if t_duration is None:
-                raise_error(ValueError, f"In the hamiltonian_simulation mode you need to work with evolution operators so please specify a time.")
-            else:
-                raise_error(TypeError, "Not implemented")
-                return sum Z @ J_input.circuit(t_duration) @Z
-
-
-
-
-### @TODO discuss with others if its not better to formulate it like this
-
-
-
-
-class doubleBracketStep:
-    def __init__(
-    self,
-    s_step: double = None,
-    d_Z: DiagonalAssociation = None,
-    mode_dbr: DoubleBracketRotationType = DoubleBracketRotationType.single_commutator,
-    mode_evolutiom_reversal: EvolutionOracleInputHamiltonianReversalType = None,
-    mode_evolution_oracle: EvolutionOracleType = EvolutionOracleType.text_strings
-):      
-        self.s_step = s_step
-        self.diagonal_association = d_Z
-        self.mode_dbr = mode_dbr #@TODO: this should allow to request gradient search or other operator optimization
-        self.mode_gci_reversal = mode_GCI_reversal
-        self.mode_dbr = mode_dbr 
-        self.mode_evolution_oracle = mode_evolution_oracle
-
-    def loss(self, step: float, look_ahead: int = 1):
-        """
-        Compute loss function distance between `look_ahead` steps.
-
-        Args:
-            step: iteration step.
-            look_ahead: number of iteration steps to compute the loss function;
-        """
-        # copy initial hamiltonian
-        h_copy = deepcopy(self.h)
-
-        for _ in range(look_ahead):
-            self.__call__(mode=self.mode, step=step)
-
-        # off_diagonal_norm's value after the steps
-        loss = self.off_diagonal_norm
-
-        # set back the initial configuration
-        self.h = h_copy
-
-        return loss
-
-    def energy_fluctuation(self, state):
-        """
-        Evaluate energy fluctuation
-
-        .. math::
-            \\Xi_{k}(\\mu) = \\sqrt{\\langle\\mu|\\hat{H}^2|\\mu\\rangle - \\langle\\mu|\\hat{H}|\\mu\\rangle^2} \\,
-
-        for a given state :math:`|\\mu\\rangle`.
-
-        Args:
-            state (np.ndarray): quantum state to be used to compute the energy fluctuation with H.
-        """
-        return self.h.energy_fluctuation(state)
-
-    @property
-    def backend(self):
-        """Get Hamiltonian's backend."""
-        return self.h0.backend
-
-    def hyperopt_step(
-        self,
-        step_min: float = 1e-5,
-        step_max: float = 1,
-        max_evals: int = 1000,
-        space: callable = None,
-        optimizer: callable = None,
-        look_ahead: int = 1,
-        verbose: bool = False,
-    ):
-        """
-        Optimize iteration step.
-
-        Args:
-            step_min: lower bound of the search grid;
-            step_max: upper bound of the search grid;
-            max_evals: maximum number of iterations done by the hyperoptimizer;
-            space: see hyperopt.hp possibilities;
-            optimizer: see hyperopt algorithms;
-            look_ahead: number of iteration steps to compute the loss function;
-            verbose: level of verbosity.
-
-        Returns:
-            (float): optimized best iteration step.
-        """
-        if space is None:
-            space = hyperopt.hp.uniform
-        if optimizer is None:
-            optimizer = hyperopt.tpe
-
-        space = space("step", step_min, step_max)
-        best = hyperopt.fmin(
-            fn=partial(self.loss, look_ahead=look_ahead),
-            space=space,
-            algo=optimizer.suggest,
-            max_evals=max_evals,
-            verbose=verbose,
-        )
-
-        return best["step"]
-
-
-class DoubleBracketIterationNumpy(DoubleBracketIteration):
-    """
-    Class implementing the Double Bracket iteration algorithm.
-    For more details, see https://arxiv.org/pdf/2206.11772.pdf
-
-    Args:
-        hamiltonian (Hamiltonian): Starting Hamiltonian;
-        mode (DoubleBracketDiagonalAssociationType): type of generator of the evolution.
-
-    Example:
-        .. testcode::
-
-            import numpy as np
-            from qibo.models.dbi.double_bracket import DoubleBracketIteration, DoubleBracketDiagonalAssociationType
-            from qibo.hamiltonians import Hamiltonian
-            from qibo.quantum_info import random_hermitian
-
-            nqubits = 4
-            h0 = random_hermitian(2**nqubits)
-            dbf = DoubleBracketIteration(Hamiltonian(nqubits=nqubits, matrix=h0))
-
-            # diagonalized matrix
-            dbf.h
-    """
-
-    def __init__(
-        self,
-        hamiltonian: Hamiltonian,
-        mode: DoubleBracketDiagonalAssociationType = DoubleBracketDiagonalAssociationType.canonical,
-    ):
-        self.h = hamiltonian
-        self.h0 = deepcopy(self.h)
-        self.mode = mode
-
-    def __call__(
-        self, step: float, mode: DoubleBracketDiagonalAssociationType = None, d: np.array = None
-    ):
-        if mode is None:
-            mode = self.mode
-
-        if mode is DoubleBracketDiagonalAssociationType.canonical:
-            operator = self.backend.calculate_matrix_exp(
-                1.0j * step,
-                self.commutator(self.diagonal_h_matrix, self.h.matrix),
-            )
-        elif mode is DoubleBracketDiagonalAssociationType.single_commutator:
-            if d is None:
-                raise_error(ValueError, f"Cannot use group_commutator with matrix {d}")
-            operator = self.backend.calculate_matrix_exp(
-                1.0j * step,
-                self.commutator(d, self.h.matrix),
-            )
-        elif mode is DoubleBracketDiagonalAssociationType.group_commutator:
-            if d is None:
-                raise_error(ValueError, f"Cannot use group_commutator with matrix {d}")
-            operator = (
-                self.h.exp(-step)
-                @ self.backend.calculate_matrix_exp(-step, d)
-                @ self.h.exp(step)
-                @ self.backend.calculate_matrix_exp(step, d)
-            )
-        operator_dagger = self.backend.cast(
-            np.matrix(self.backend.to_numpy(operator)).getH()
-        )
-        self.h.matrix = operator @ self.h.matrix @ operator_dagger
-
-    @staticmethod
-    def commutator(a, b):
-        """Compute commutator between two arrays."""
-        return a @ b - b @ a
-
-    @property
-    def diagonal_h_matrix(self):
-        """Diagonal H matrix."""
-        return self.backend.cast(np.diag(np.diag(self.backend.to_numpy(self.h.matrix))))
-
-    @property
-    def off_diag_h(self):
-        return self.h.matrix - self.diagonal_h_matrix
-
-    @property
-    def off_diagonal_norm(self):
-        """Norm of off-diagonal part of H matrix."""
-        off_diag_h_dag = self.backend.cast(
-            np.matrix(self.backend.to_numpy(self.off_diag_h)).getH()
-        )
-        return np.real(
-            np.trace(self.backend.to_numpy(off_diag_h_dag @ self.off_diag_h))
-        )
-
-
